# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnonymousSessionRequest": {
        "properties": {
          "host_id": {
            "pattern": "^AMULAI-HOST-[0-9a-f]{8}$",
            "title": "Host Id",
            "type": "string"
          },
          "locale": {
            "enum": [
              "en",
              "hi",
              "gu"
            ],
            "title": "Locale",
            "type": "string"
          }
        },
        "required": [
          "host_id",
          "locale"
        ],
        "title": "AnonymousSessionRequest",
        "type": "object"
      },
      "AnonymousSessionResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "features": {
            "items": {
              "enum": [
                "advisory_chat",
                "voice_input"
              ],
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "access_token",
          "expires_in",
          "features"
        ],
        "title": "AnonymousSessionResponse",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "locale": {
            "enum": [
              "en",
              "hi",
              "gu"
            ],
            "title": "Locale",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "text",
          "locale"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PublicHostConfig": {
        "properties": {
          "default_locale": {
            "enum": [
              "en",
              "hi",
              "gu"
            ],
            "title": "Default Locale",
            "type": "string"
          },
          "features": {
            "items": {
              "enum": [
                "advisory_chat",
                "voice_input"
              ],
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "host_id": {
            "title": "Host Id",
            "type": "string"
          },
          "locales": {
            "items": {
              "enum": [
                "en",
                "hi",
                "gu"
              ],
              "type": "string"
            },
            "title": "Locales",
            "type": "array"
          },
          "partner_name": {
            "title": "Partner Name",
            "type": "string"
          }
        },
        "required": [
          "host_id",
          "partner_name",
          "features",
          "locales",
          "default_locale"
        ],
        "title": "PublicHostConfig",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Amul AI Widget BFF",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/chat/stream": {
      "post": {
        "operationId": "chat_stream_api_v1_chat_stream_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream"
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Api V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/v1/hosts/{host_id}/config": {
      "get": {
        "operationId": "host_config_api_v1_hosts__host_id__config_get",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicHostConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Host Config"
      }
    },
    "/api/v1/sessions/anonymous": {
      "post": {
        "operationId": "anonymous_session_api_v1_sessions_anonymous_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymousSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnonymousSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anonymous Session"
      }
    },
    "/embed/{host_id}": {
      "get": {
        "operationId": "embed_embed__host_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Embed"
      }
    }
  }
}